requirements.txt
flask>=3.0
pandas>=2.2.0
numpy>=1.26.0
SQLAlchemy>=2.0.25
discord.py>=2.3.2
python-dotenv>=1.0.1
google-genai==0.7.0
requests>=2.0.0
matplotlib>=3.8.0
armoriq-sdk>=0.3.0
